refactoring data layer clean architecture model
