+++ app/src/main/java/com/example/englishwords/data/sources/room/mapper/RoomMapper.kt
@@ -1,7 +1,7 @@
 package com.example.englishwords.data.sources.room.mapper
 
-import com.example.englishwords.data.sources.Mapper
-import com.example.englishwords.data.sources.model.WordDbModel
+import com.example.englishwords.data.Mapper
+import com.example.englishwords.data.model.WordDbModel
 import com.example.englishwords.domain.models.Word
 
 class RoomMapper: Mapper {
