+++ app/src/main/java/com/example/englishwords/viewmodel/LetterViewModel.kt
@@ -2,7 +2,7 @@ package com.example.englishwords.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.englishwords.data.sources.model.WordDbModel
+import com.example.englishwords.data.model.WordDbModel
 import com.example.englishwords.data.sources.room.db.DbManager
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
